Continue to overhaul the int/long int, index number and sscanf handling.
- Replace sscanf() with strtol() almost everywhere. Better error
  checking for me, and no different behaviors on Linux/*BSD.
- Try to use long int for indices, and variables that correspond to indices across the source.
- Some variable renaming.
- More range checking when dealing with user input indices.

+++ cmd_clist.c
@@ -36,7 +36,7 @@ cmd_clist(const char *e_line, command *commands)
 	xmlNodePtr	db_node = NULL;
 	xmlChar		*name = NULL;
 
-	int		idx = 0;
+	long int	idx = 0;
 
 
 	db_node = keychain->parent->children;
@@ -47,7 +47,7 @@ cmd_clist(const char *e_line, command *commands)
 	while (db_node) {
 		if (db_node->type == XML_ELEMENT_NODE) {	/* we only care about ELEMENT nodes */
 			name = xmlGetProp(db_node, BAD_CAST "name");
-			printf("%d. %s\n", idx++, name);
+			printf("%ld. %s\n", idx++, name);
 			xmlFree(name); name = NULL;
 		}
